backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyjwt==2.8.0
python-multipart==0.0.9
